+++ java/src/org/chromium/android_webview/AwPicture.java
@@ -8,7 +8,6 @@ import android.graphics.Canvas;
 import android.graphics.Picture;
 import android.graphics.Rect;
 
-import org.chromium.base.CalledByNative;
 import org.chromium.base.JNINamespace;
 import org.chromium.content.common.CleanupReference;
 
@@ -19,13 +18,13 @@ import java.io.OutputStream;
 @JNINamespace("android_webview")
 class AwPicture extends Picture {
 
-    private int mNativeAwPicture;
+    private long mNativeAwPicture;
 
     // There is no explicit destroy method on Picture base-class, so cleanup is always
     // handled via the CleanupReference.
     private static final class DestroyRunnable implements Runnable {
-        private int mNativeAwPicture;
-        private DestroyRunnable(int nativeAwPicture) {
+        private long mNativeAwPicture;
+        private DestroyRunnable(long nativeAwPicture) {
             mNativeAwPicture = nativeAwPicture;
         }
         @Override
@@ -40,7 +39,7 @@ class AwPicture extends Picture {
      * @param nativeAwPicture is an instance of the AwPicture native class. Ownership is
      *                        taken by this java instance.
      */
-    AwPicture(int nativeAwPicture) {
+    AwPicture(long nativeAwPicture) {
         mNativeAwPicture = nativeAwPicture;
         mCleanupReference = new CleanupReference(this, new DestroyRunnable(nativeAwPicture));
     }
@@ -86,10 +85,10 @@ class AwPicture extends Picture {
         throw new IllegalStateException("Unsupported in AwPicture");
     }
 
-    private static native void nativeDestroy(int nativeAwPicture);
-    private native int nativeGetWidth(int nativeAwPicture);
-    private native int nativeGetHeight(int nativeAwPicture);
-    private native void nativeDraw(int nativeAwPicture, Canvas canvas,
+    private static native void nativeDestroy(long nativeAwPicture);
+    private native int nativeGetWidth(long nativeAwPicture);
+    private native int nativeGetHeight(long nativeAwPicture);
+    private native void nativeDraw(long nativeAwPicture, Canvas canvas,
             int left, int top, int right, int bottom);
 }
