+++ java/src/org/chromium/base/SystemMessageHandler.java
@@ -1,4 +1,4 @@
-// Copyright (c) 2012 The Chromium Authors. All rights reserved.
+// Copyright 2012 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -6,9 +6,6 @@ package org.chromium.base;
 
 import android.os.Handler;
 import android.os.Message;
-import android.os.SystemClock;
-
-import java.util.concurrent.atomic.AtomicBoolean;
 
 class SystemMessageHandler extends Handler {
 
@@ -16,9 +13,9 @@ class SystemMessageHandler extends Handler {
     private static final int DELAYED_TIMER_MESSAGE = 2;
 
     // Native class pointer set by the constructor of the SharedClient native class.
-    private int mMessagePumpDelegateNative = 0;
+    private long mMessagePumpDelegateNative = 0;
 
-    private SystemMessageHandler(int messagePumpDelegateNative) {
+    private SystemMessageHandler(long messagePumpDelegateNative) {
         mMessagePumpDelegateNative = messagePumpDelegateNative;
     }
 
@@ -47,9 +44,9 @@ class SystemMessageHandler extends Handler {
     }
 
     @CalledByNative
-    private static SystemMessageHandler create(int messagePumpDelegateNative) {
+    private static SystemMessageHandler create(long messagePumpDelegateNative) {
         return new SystemMessageHandler(messagePumpDelegateNative);
     }
 
-    private native void nativeDoRunLoopOnce(int messagePumpDelegateNative);
+    private native void nativeDoRunLoopOnce(long messagePumpDelegateNative);
 }
